충돌 해결

+++ frontend/src/page/post/articles/ArticleModify.vue
@@ -269,13 +269,15 @@ let longs = ''
         Bucket: this.albumBucketName
       }
     })
-
-      s3.deleteObject({ Key: photoKey }, (err) => {
-      if (err) {
-    
-      }
       
-    });
+      if(photoKey != null){
+        s3.deleteObject({ Key: photoKey }, (err) => {
+        if (err) {
+      
+        }
+        
+      });
+      }
       }
   }
 }
